+++ app/api/admin/ai-usage/route.ts
@@ -1,25 +1,10 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { adminAuth, adminDb } from '@/lib/firebase-admin';
+import { adminDb } from '@/lib/firebase-admin';
+import { authorizeAdmin } from '@/app/api/admin/AuthorizeAPI';
 
 function getTodayString(): string {
   return new Date().toISOString().split('T')[0];
-}
-
-async function authorizeAdmin(req: NextRequest): Promise<{ ok: true; uid: string } | { ok: false }> {
-  const authHeader = req.headers.get('authorization');
-  if (!authHeader || !authHeader.startsWith('Bearer ')) return { ok: false };
-
-  const idToken = authHeader.slice('Bearer '.length);
-  try {
-    const decoded = await adminAuth.verifyIdToken(idToken);
-    if (decoded?.uid && decoded.admin === true) {
-      return { ok: true, uid: decoded.uid };
-    }
-    return { ok: false };
-  } catch {
-    return { ok: false };
-  }
 }
 
 export async function GET(req: NextRequest) {
