+++ app/api/admin/embeddings/route.ts
@@ -1,24 +1,8 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { adminAuth } from '@/lib/firebase-admin';
 import { fetchCourses } from '@/lib/courses';
 import { generateAndStoreCourseEmbedding, getEmbeddingCount } from '@/lib/ai/vector-store';
-
-async function authorizeSuperAdmin(req: NextRequest): Promise<{ ok: true; uid: string } | { ok: false }> {
-  const authHeader = req.headers.get('authorization');
-  if (!authHeader || !authHeader.startsWith('Bearer ')) return { ok: false };
-
-  const idToken = authHeader.slice('Bearer '.length);
-  try {
-    const decoded = await adminAuth.verifyIdToken(idToken);
-    if (decoded?.uid && decoded.superAdmin === true) {
-      return { ok: true, uid: decoded.uid };
-    }
-    return { ok: false };
-  } catch {
-    return { ok: false };
-  }
-}
+import { authorizeSuperAdmin } from '@/app/api/admin/AuthorizeAPI';
 
 export async function GET(req: NextRequest) {
   try {
