+++ app/api/admin/parse-requirements/route.ts
@@ -1,24 +1,9 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { adminAuth, adminDb } from '@/lib/firebase-admin';
+import { adminDb } from '@/lib/firebase-admin';
 import { fetchCourses } from '@/lib/courses';
 import { parseRequirements } from '@/lib/prerequisites/parser';
-
-async function authorizeAdmin(req: NextRequest): Promise<{ ok: true; uid: string } | { ok: false }> {
-    const authHeader = req.headers.get('authorization');
-    if (!authHeader || !authHeader.startsWith('Bearer ')) return { ok: false };
-
-    const idToken = authHeader.slice('Bearer '.length);
-    try {
-        const decoded = await adminAuth.verifyIdToken(idToken);
-        if (decoded?.uid && (decoded.admin === true || decoded.superAdmin === true)) {
-            return { ok: true, uid: decoded.uid };
-        }
-        return { ok: false };
-    } catch {
-        return { ok: false };
-    }
-}
+import { authorizeAdmin } from '@/app/api/admin/AuthorizeAPI';
 
 /**
  * POST /api/admin/parse-requirements
