+++ app/api/courses/embedding-map/route.ts
@@ -2,6 +2,7 @@ import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 import { adminDb } from '@/lib/firebase-admin';
 import { projectEmbeddings, type ProjectionAlgorithm } from '@/lib/ai/projection';
+import { authorizeAdmin } from '@/app/api/admin/AuthorizeAPI';
 
 export interface EmbeddingPoint {
     courseId: string;
@@ -31,6 +32,10 @@ function normalizeLevel(lvl: unknown): string {
 }
 
 export async function GET(req: NextRequest) {
+    const auth = await authorizeAdmin(req);
+    if (!auth.ok) {
+        return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
+    }
     try {
         const url = new URL(req.url);
         const dims = (parseInt(url.searchParams.get('dimensions') || '2') === 3 ? 3 : 2) as 2 | 3;
